Create ComputerChoice randomizer and game functions
We create the game functions using Object Oriented
Programming principles. We first create a function
which return a random choice from an array. Then we
create a function which compares the choice with user
input and returns a string depending on the result.
Lastly, we create a PlayGame function which loops over
our previous function, basically playing "the game"
a few times. At the same time, 2 helper variables are
used to keep track of the score (which we are tracking
using the .includes() method since our return values
are strings) and prints to the console a string based
on a comparasion between the values :D